requirements.txt
# Plot Analysis Generator - Requirements
# Install with: pip install -r requirements.txt
# Updated to match Pipeline v2.0

# ==============================================================================
# Core Geospatial
# ==============================================================================
shapely>=2.0.0
geopandas>=0.14.0
pyproj>=3.6.0
fiona>=1.9.0

# ==============================================================================
# OSM Data Collection
# ==============================================================================
overpy>=0.7
osmnx>=1.9.0

# ==============================================================================
# HTTP/API Requests
# ==============================================================================
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.26.0

# ==============================================================================
# Imagery & Terrain (Mapbox + Rasters)
# ==============================================================================
rasterio>=1.3.0
rio-tiler>=6.0.0  # For Mapbox Terrain-RGB tiles
mercantile>=1.2.1  # Tile math
pillow>=10.0.0

# ==============================================================================
# Solar / Shadow Analysis
# ==============================================================================
pvlib>=0.10.0
ephem>=4.1.0  # Alternative to pyephem

# ==============================================================================
# Vegetation Segmentation (Optional - for SAM)
# ==============================================================================
# segment-anything>=1.0  # Install separately if using SAM
# torch>=2.0.0  # Required for SAM

# ==============================================================================
# Data Processing
# ==============================================================================
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
scikit-image>=0.22.0

# ==============================================================================
# Data Models & Validation
# ==============================================================================
pydantic>=2.5.0
orjson>=3.9.0

# ==============================================================================
# CLI & Utilities
# ==============================================================================
click>=8.1.0
tqdm>=4.66.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0  # For loading .env files

# ==============================================================================
# Logging
# ==============================================================================
loguru>=0.7.0

# ==============================================================================
# Visualization (Optional)
# ==============================================================================
matplotlib>=3.8.0

# ==============================================================================
# Database (Optional - for production)
# ==============================================================================
# pymongo>=4.6.0  # For MongoDB storage
# motor>=3.3.0  # Async MongoDB
